src/data/kalshi_fetcher.py
"""
Fetcher for Kalshi contract prices and market data.

TODO: Implement data fetching from Kalshi API.
This module should handle:
  - Authentication via KALSHI_API_KEY from environment
  - Fetching contract prices (historical and real-time)
  - Mapping NBA games to Kalshi market IDs
"""

import os
from dotenv import load_dotenv
from src.utils.logger import get_logger

logger = get_logger(__name__)

load_dotenv()
KALSHI_API_KEY = os.getenv("KALSHI_API_KEY")


def fetch_contracts(market_id, date_from=None, date_to=None):
    """
    Fetch contract price history for a given market.

    Args:
        market_id (str): Kalshi market ID
        date_from (str, optional): Start date in YYYY-MM-DD format
        date_to (str, optional): End date in YYYY-MM-DD format

    Returns:
        pd.DataFrame: Price history with columns [timestamp, price, volume, ...]
    """
    # TODO: Implement Kalshi API calls
    if not KALSHI_API_KEY:
        logger.warning("KALSHI_API_KEY not set in environment")
    logger.info(f"Fetching contracts for market {market_id}...")
    raise NotImplementedError("fetch_contracts() not yet implemented")


def fetch_current_price(market_id):
    """
    Fetch current price for a contract.

    Args:
        market_id (str): Kalshi market ID

    Returns:
        float: Current contract price (market-implied probability)
    """
    # TODO: Implement Kalshi API calls
    logger.info(f"Fetching current price for market {market_id}...")
    raise NotImplementedError("fetch_current_price() not yet implemented")
