requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0
nba_api>=1.4
requests>=2.31
python-dotenv>=1.0
pyyaml>=6.0
jupyterlab>=4.0
pytest>=8.0
